+++ pages/user/[id]/profile.vue
@@ -1,4 +1,5 @@
 <template>
+    <Title>User profile for {{ user.meta?.displayName }}</Title>
     <div v-if="user.props?.id" class="surface-card p-4 border-1 surface-border border-round flex-auto">
         <div class="text-900 font-semibold text-lg mt-3">User profile for {{ user.meta?.displayName }}</div>