+++ pages/user/[id]/reset-password.vue
@@ -1,10 +1,14 @@
 <template>
+    <Title>Create an Account for {{ config.public?.site?.name }}</Title>
     <div class="w-full lg:w-7 mx-auto">
         <div class="text-center mb-5">
             <img :src="config.public.site.logoUrl" :alt="config.public.site.name" height="50" class="mb-3">
+            <h1 class="text-900 text-3xl font-medium mb-3 mt-0">Reset Password</h1>
+            <span class="text-600 font-medium line-height-3">Remember your account?</span>
+            <NuxtLink to="/user/login" class="font-medium no-underline ml-2 text-blue-500 cursor-pointer">Sign in
+            </NuxtLink>
         </div>
-        <div class="surface-card p-4 border-1 surface-border border-round ">
-            <h1 class="mt-0">Reset Passsword</h1>
+        <PanelZone>
             <Form :send="resetPassword">
                 <div class="flex gap-5 flex-column-reverse md:flex-row">
                     <div class="flex-auto">
@@ -25,7 +29,7 @@
 
                 </div>
             </Form>
-        </div>
+        </PanelZone>
     </div>
 </template>