+++ pages/user/logout.vue
@@ -1,4 +1,5 @@
 <template>
+    <Title>Logged out of {{ config.public?.site?.name }}</Title>
     <p>You are now logged out.</p>
 
     <NuxtLink to="/" v-slot="{ navigate }" custom>
@@ -12,6 +13,7 @@ import { sleep } from '#ving/utils/sleep.mjs';
 definePageMeta({
     middleware: 'auth'
 });
+const config = useRuntimeConfig();
 const currentUser = useCurrentUser();
 await currentUser.logout();
 await sleep(1000 * 5);