+++ pages/user/must-verify-email.vue
@@ -1,12 +1,13 @@
 <template>
+    <Title>Must verify email for {{ config.public?.site?.name }}</Title>
     <div class="w-full lg:w-7 mx-auto">
         <div class="text-center mb-5">
             <img :src="config.public.site.logoUrl" :alt="config.public.site.name" height="50" class="mb-3">
         </div>
-        <div class="surface-card p-4 border-1 surface-border border-round ">
+        <PanelZone>
             You cannot continue to this content until you have verified your email address. We've sent you an email with a
             link to verify. Please check your email to continue.
-        </div>
+        </PanelZone>
     </div>
 </template>